+++ misc/python/materialize/scalability/endpoint/endpoint.py
