+++ misc/python/materialize/scalability/operation/operation_data.py
