+++ misc/python/materialize/scalability/result/comparison_outcome.py
@@ -14,8 +14,8 @@ from materialize.scalability.df.df_totals import (
     DfTotalsExtended,
     concat_df_totals_extended,
 )
-from materialize.scalability.endpoint import Endpoint
-from materialize.scalability.scalability_change import (
+from materialize.scalability.endpoint.endpoint import Endpoint
+from materialize.scalability.result.scalability_change import (
     Regression,
     ScalabilityChange,
     ScalabilityImprovement,
