+++ misc/python/materialize/scalability/workload/workload_markers.py
@@ -6,9 +6,7 @@
 # As of the Change Date specified in that file, in accordance with
 # the Business Source License, use of this software will be governed
 # by the Apache License, Version 2.0.
-
-
-from materialize.scalability.workload import Workload
+from materialize.scalability.workload.workload import Workload
 
 
 class WorkloadMarker(Workload):
