+++ misc/python/materialize/scalability/workload/workloads/connection_workloads.py
@@ -6,21 +6,18 @@
 # As of the Change Date specified in that file, in accordance with
 # the Business Source License, use of this software will be governed
 # by the Apache License, Version 2.0.
-
-
-from materialize.scalability.operation import Operation, OperationChainWithDataExchange
-from materialize.scalability.operation_data import OperationData
-from materialize.scalability.operations import (
+from materialize.scalability.operation.operation_data import OperationData
+from materialize.scalability.operation.operations.operations import (
     Connect,
     Disconnect,
     SelectOne,
 )
-from materialize.scalability.workload import (
-    WorkloadWithContext,
+from materialize.scalability.operation.scalability_operation import (
+    Operation,
+    OperationChainWithDataExchange,
 )
-from materialize.scalability.workload_markers import (
-    ConnectionWorkload,
-)
+from materialize.scalability.workload.workload import WorkloadWithContext
+from materialize.scalability.workload.workload_markers import ConnectionWorkload
 
 
 class EstablishConnectionWorkload(WorkloadWithContext, ConnectionWorkload):
