+++ test/scalability/mzcompose.py
@@ -26,14 +26,11 @@ from materialize.mzcompose.services.materialized import Materialized
 from materialize.mzcompose.services.mz import Mz
 from materialize.mzcompose.services.postgres import Postgres
 from materialize.mzcompose.test_result import FailedTestExecutionError
-from materialize.scalability.benchmark_config import BenchmarkConfiguration
-from materialize.scalability.benchmark_executor import BenchmarkExecutor
-from materialize.scalability.benchmark_result import BenchmarkResult
-from materialize.scalability.comparison_outcome import ComparisonOutcome
+from materialize.scalability.config.benchmark_config import BenchmarkConfiguration
 from materialize.scalability.df import df_totals_cols
 from materialize.scalability.df.df_totals import DfTotalsExtended
-from materialize.scalability.endpoint import Endpoint
-from materialize.scalability.endpoints import (
+from materialize.scalability.endpoint.endpoint import Endpoint
+from materialize.scalability.endpoint.endpoints import (
     TARGET_HEAD,
     TARGET_MATERIALIZE_LOCAL,
     TARGET_MATERIALIZE_REMOTE,
@@ -44,25 +41,26 @@ from materialize.scalability.endpoints import (
     PostgresContainer,
     endpoint_name_to_description,
 )
+from materialize.scalability.executor.benchmark_executor import BenchmarkExecutor
 from materialize.scalability.io import paths
 from materialize.scalability.plot.plot import (
     plot_duration_by_connections_for_workload,
     plot_duration_by_endpoints_for_workload,
     plot_tps_per_connections,
 )
-from materialize.scalability.regression_assessment import RegressionAssessment
-from materialize.scalability.result_analyzer import ResultAnalyzer
-from materialize.scalability.result_analyzers import DefaultResultAnalyzer
+from materialize.scalability.result.benchmark_result import BenchmarkResult
+from materialize.scalability.result.comparison_outcome import ComparisonOutcome
+from materialize.scalability.result.regression_assessment import RegressionAssessment
+from materialize.scalability.result.result_analyzer import ResultAnalyzer
+from materialize.scalability.result.result_analyzers import DefaultResultAnalyzer
 from materialize.scalability.scalability_versioning import SCALABILITY_FRAMEWORK_VERSION
-from materialize.scalability.schema import Schema, TransactionIsolation
-from materialize.scalability.workload import (
-    Workload,
-)
-from materialize.scalability.workload_markers import WorkloadMarker
-from materialize.scalability.workloads.connection_workloads import *  # noqa: F401 F403
-from materialize.scalability.workloads.ddl_workloads import *  # noqa: F401 F403
-from materialize.scalability.workloads.dml_dql_workloads import *  # noqa: F401 F403
-from materialize.scalability.workloads.self_test_workloads import *  # noqa: F401 F403
+from materialize.scalability.schema.schema import Schema, TransactionIsolation
+from materialize.scalability.workload.workload import Workload
+from materialize.scalability.workload.workload_markers import WorkloadMarker
+from materialize.scalability.workload.workloads.connection_workloads import *  # noqa: F401 F403
+from materialize.scalability.workload.workloads.ddl_workloads import *  # noqa: F401 F403
+from materialize.scalability.workload.workloads.dml_dql_workloads import *  # noqa: F401 F403
+from materialize.scalability.workload.workloads.self_test_workloads import *  # noqa: F401 F403
 from materialize.test_analytics.config.test_analytics_db_config import (
     create_test_analytics_config,
 )
